Part 4, with first bit of coding
This code is awful, terrible and bad. But this is all exploratory while
we get into the vibe. We're discussing a lot as we go, and what's clear
is that we need to refine the process a little.

As time goes by this stuff won't take as long to construct.